+++ HISTORY.rst
@@ -22,6 +22,7 @@ History
 - Add file actions to detail view
 - Show dialogs pointing the user to the Rana HCC when creating a model, uploading a revision and starting a simulation
 - Only show 3Di organisation dropdown if more than 1 organisation is available per tenant
+- Add confirmation for delete
 
 
 1.1.20 (2025-11-18)